policy: Add an integration test to exercise the gRPC API
This change adds a simple policy integration test that exercises an
installed Linkerd control plane to verify that the gRPC API responds to
resource changes in the control plane.

This test works by establishing a port-forward to the gRPC server,
creating resources, and verifying that an API watch updates as expected.

Signed-off-by: Oliver Gould <ver@buoyant.io>