Repair endpoints for headless mirror services (#8119)
This is a follow-up to #8066 which was a bad merge; it used
`createOrUpdateEndpoints` which was removed in favor of `createMirrorEndpoints`
and `updateMirrorEndpoints`. dfb79fda7c03dfca295856478b8cd3c3d3442160 fixed this
incorrect usage.

---

#8022 fixed an issue where if the gateway for a target cluster was down, the
endpoints of mirror services still indicated their addresses (the gateway in
this case) were ready. This was not fixed for _headless_ mirror services and a
comment was left explaining the issue
[here](https://github.com/linkerd/linkerd2/blob/9829f486a0cdb49b329156259943e33436102ecc/multicluster/service-mirror/cluster_watcher_headless.go#L380-L386).

This fixes that last issue and ties gateway liveness to the endpoints for
headless mirror services now as well. This involved changes to the two sets of
endpoints created for each headless mirror service.

1. The first set of endpoints point to auxiliary services that actually target
   the gateway addresses. Therefore, these addresses should always be ready and
   we skip the `gatewayAlive` check—going through just an API `Create`.
2. The second set of endpoints belong to these auxiliary services and target the
   gateway addresses. These are the endpoints that are created through
   `createOrUpdateEndpoints` — ensuring that `gatewayAlive` is checked — and are
   the endpoints that we care about repairing. `repairEndpoints` now attempts to
   repair the endpoints that belong to these auxiliary services by defaulting
   their addresses to ready which is then checked by `createOrUpdateEndpoints`.

### Testing

No tests were added because we're already properly testing this. If we just
replace the `Endpoints(...).Update(...)` with `createOrUpdateEndpoints`,
integration tests start failing because we never attempt to repair the endpoints
created as not ready.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>