policy-test: Retry waiting for lock deletion
We sometimes see timeouts in the policy e2e tests, as it seems `kubectl
wait` may not be notified of lock deletion.

To be more resilient to this situation, this change updates curl pods to
run `kubectl wait` with a shorter timeout with retries. The overall
timeout (2 minutes) remains unchanged.

Signed-off-by: Oliver Gould <ver@buoyant.io>