ci: Test policy on the minimum k8s version (v1.20) (#8252)
The policy tests only run on the most recent Kubernetes version. This
change updates the policy test workflow to use a matrix to run these
tests on the minimum-supported kubernetes version.

In order for the tests to pass on k8s v1.20, we need to block resource
creation on service account availability. Otherwise, pods may fail due
to the serviceaccount not being available.

The policy test runner is changed to nextest so that we can build the
tests once and share the artifacts in all instances of the cluster tests.

Signed-off-by: Oliver Gould <ver@buoyant.io>