policy-test: Explicitly wait for init containers (#8272)
The policy integration tests sometimes fail in CI due to timeouts
waiting for a curl pod's exit code. It's not really clear why this is
happening.

This change separates waiting for the init containers from waiting for
the exit code so we can log that situation explicitly. Furthermore,
logging has been enhanced to include information about init containers.

`curl` invocations now use a timeout and retries to help limit flakiness.
Also, `curl-lock` deletions now wait for the resource to be fully removed.

Signed-off-by: Oliver Gould <ver@buoyant.io>