Introduce fs watch for cni installer
Our CNI installer script is prone to race conditions, especially when a
node is rebooted, or restarted. Order of configuration should not matter
and our CNI plugin should attach to other plugins (i.e chain to them) or
run standalone when applicable. In order to be more flexible, we
introduce a filesystem watcher through inotifywait to react to changes
in the cni config directory.

To make the script easier to work with, I have also refactored the code
into smaller, reusable functions.

Signed-off-by: Matei David <matei@buoyant.io>