Add SHA-based comparison for fs changes
This commit refactors most of the script, once again, to get rid of
unnecessarily overridden ENV variables. In the script now, we opt to use
mostly `local` variables inside function bodies. Rather than maintain a
CONFIG_PATH env variable, we do most of our comparisons based on SHAs.

Linkerd's CNI plugin should append configuration when at least one other
file exists, but if multiple files exist, the CNI plugin should not have
to make a decision on whether thats the current file to append itself
to. As a result, most of the logic in this commit revolves around the
assumption that whatever file we detect has been created should be
injected with Linkerd's config -- the rest is up to the host.

Whenever another file exists, Linkerd's CNI plugin will check whether
the default configuration (i.e when _we_ run in interface mode) still
exists, and if it does, it is cleaned up.

Whenever we get a DELETE event, we only re-install if no other
configuration files exist. We re-install during CREATE events only if
the SHA has changed.

Signed-off-by: Matei David <matei@buoyant.io>