release: Don't use a dedicated account to push images
We currently push images using a dedicated GitHub login/token. This is
unnecessary if we give the workflow sufficient permissions to push
packages.

This change updates the release workflow to use `docker/login-action` to
log into GHCR with the action's local secret so that pushes no longer
rely on a dedicated bot account.

Signed-off-by: Oliver Gould <ver@buoyant.io>