Introduce proto retries to Service Profiles
Currently, proto Service Profiles are generated without setting the
IsRetryable field which limits the resiliency you get out of the box
as requests are never retried unless you manually update it.

This change sets the IsRetryable field to true if
idempotency_level is either IDEMPOTENT or NO_SIDE_EFFECTS.

The new functionality is tested by extending the existing unit tests.

Fixes: #8472

Signed-off-by: Martin Anker Have <mah@lunar.app>