Fix multicluster headless svc mirroring (#8474)
* Fix multicluster headless svc mirroring

Fixes #8469

** Note this needs to be back-ported into 2.11.x **

When labeling a headless service for exporting, the service-mirror controller was ignoring the change because `isExportedEndpoints()` was checked on the _old_ version of the EP update (which doesn't have the label).

This change checks on both the old and new version of the updated EP (and so I had to move some of the stuff out of `isExportedEndpoints()` into the calling sites, to avoid dupped log entries when calling it).

In order to test this, I removed the `mirror.linkerd.io/exported` label from the nginx SS and instead label it in the test after the SS persists.

* Consolidated isExportedEndpoints() and isExportedService() into isExported(), which uses the label configured in the Link

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>