Add safe-to-evict annotation to control plane components (#8524)
Fixes #4067

We add the `cluster-autoscaler.kubernetes.io/safe-to-evict: "true"` annotation to the Linkerd control plane components.  This annotation tells the cluster autoscaler that even though the control plane components have volume mounts, it is okay to evict them (subject to pod disruption constraints).  This is because we only use the volume as temporary storage for certificates and do not need to persist that data.

Signed-off-by: Alex Leong <alex@buoyant.io>