Set targetCluster label even when serviceFqn is not set (#8542)
Fixes #8134

The multicluster probe-gateway service (which is used by the service-mirror controller to send health probes to the remote gateway) has `mirror.linkerd.io/cluster-name` label but it does not have a `mirror.linkerd.io/remote-svc-fq-name` annotation.  This makes sense because the probe-gateway service does not correspond to any target service on the remote cluster and instead targets the gateway itself.

We update the logic in the destination controller so that we can still add the `dst_target_cluster` metric label when only the `mirror.linkerd.io/cluster-name` label is present but not the `mirror.linkerd.io/remote-svc-fq-name` annotation.  This allows us to have the `dst_target_cluster` metric label for service-mirror controller probe traffic.

Signed-off-by: Alex Leong <alex@buoyant.io>