Bump actions/cache from 3.0.2 to 3.0.3
This change supersedes the PR submitted by dependabot to update
actions/cache from 3.0.2 to 3.0.3. The dependency bump PR does not
change the version for the docker-build action that we use to build
images for our control plane components.

This change updates all references to actions/cache to use 3.0.3 (both
the release workflow, and the action).

Signed-off-by: Matei David <matei@buoyant.io>