Use service-mirror's /metrics to check gateway liveness (#8599)
Closes #8505.

The `check` command now uses the `/metrics` endpoint for each service-mirror that corresponds to a link. Using these metrics, `check` can now retrieve gateway liveness without using the `Gateways` API.

The `Gateways` API should be able to be deprecated now in a follow-up; I'm not including it as part of this change and they are separate enough changes.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>
Signed-off-by: Dmitrii Ermakov <demonihin@gmail.com>