Remove grpc-over-http (#8625)
For historical reasons, the linkerd-viz metrics API is served on a grpc-over-http protocol.  This was necessary when these API requests were being exposed through the kubernetes API.  Now we expose this API via a port-forward and the grpc-over-http protocol is no longer necessary.

We remove this protocol and instead have the metrics api simply serve plain gRPC.  We correspondingly update clients of this API to make regular gRPC requests.

Signed-off-by: Alex Leong <alex@buoyant.io>
Signed-off-by: Dmitrii Ermakov <demonihin@gmail.com>