ci: Add re-usable actions to build multiarch rust code (#8587)
In the release process, our policy-controller image cannot make use of
the typical multiarch docker build due to missing tooling for 32-bit arm
hosts. Currently, the policy-controller is cross-built from an amd64
host. In order to allow for easier integration of new Rust code into the
linkerd2 repo, and in order to promote DRY, this change splits up the
Rust docker build job steps into a composite, reusable action.

To build the policy-controller, we first build each image for its
intended target in parallel (through a matrix build), after which we
create a unified docker manifest for the image. This change pulls logic
from two respective jobs into separate actions: `rust-docker-build` and
`rust-docker-manifest`, each with a set of required inputs, such as
component name and where applicable the code path and arch.

Signed-off-by: Matei David <matei@buoyant.io>