Add DOCKER_BUILDX_CACHE path to release step
Currently, when rust code is built, we use a GitHub action that calls
the relevant scripts and logic to build images. The last step when
building an image in the release workflow is to prune the cache layers.

The script that prunes layers expects to have an environment variable
set with the path of the docker cache. During a refactor, we changed how
env variables are set for rust builds and the environment variable is no
longer set. This change adds the cache path to the prune step; missing
path results in errors when attempting to read from an inexistent file.

Signed-off-by: Matei David <matei@buoyant.io>