Add missing registry override in release workflow
Current release build is failing; normally, in the release process, we
override the docker registry that the default build action (for Go
components) is using. The default value the action assumes for the
registry is `cr.l5d.io`, but when releasing and pushing, we use GitHub's
registry (ghcr.io).

The override seems to have been accidentally removed which fails the
release process; images will be pushed to cr.l5d.io which will reply
with a 405 (Method Not Allowed). This change adds the override back in.

Signed-off-by: Matei David <matei@buoyant.io>