Fix Docker runtime check to happen only during install (#8667)
Closes #8583

Even though #7468 removed the Docker container runtime check from `linkerd check --pre` to `linkerd install` runtime error, we still do a dry run of the installation so that we can render the control plane manifests. Therefore, we still hit this check which results in not being able to run `linkerd check --pre` when nodes are using the Docker container runtime. This fixes the issue by introducing a `dryRun` flag that we check beforehand.

```shell
❯ kubectl get nodes docker-desktop -o jsonpath='{.status.nodeInfo.containerRuntimeVersion}'
docker://20.10.16

❯ bin/linkerd check --pre
Linkerd core checks
===================
...
```

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>