Allow disabling linkerd-await on control plane pods
In some circumstances, the `lifecycle.postStart` hook can cause the
`linkerd-proxy` container to get stuck waiting for identity verification.
After the `linkerd-await` timeout, the container will be restarted and the
proxy starts without further incident. The `linkerd-control-plane` helm chart
currently has a way to disable the lifecycle hook for injected proxies, but not
for proxies on the control plane pods.

This commit updates the use of the `proxy.await` value in the `linkerd-control-plane`
chart to disable the postStart lifecycle hook on the destination and proxy-injector pods.

The linkerd-control-plane chart was templated, setting
`proxy.await` to `true` and `false`, verifying that the
postStart lifecycle hook was either present or absent depending on the
`proxy.await` value.

Fixes #8738

Signed-off-by: Jacob Lambert <calrisian777@gmail.com>