Add correct match properties to link CRD to fix mirroring (#8734)
The Link created by `linkerd multicluster link` has a match expression that checks for the existance of the `mirror.linkerd.io/exported` label on services; if the label exists then the service is mirrored. This is not always correct because if `mirror.linkerd.io/exported: false`, the service mirror still mirrors the service even though it should not.

The fix for this was not related to label matching — it is already handled correctly. What needed to be fixed was the actual Link CRD to allow for `matchLabels` and `matchExpressions.values`.

The default `matchLabel` is now `mirror.linkerd.io/exported: "true"` as changed in `newLinkOptionsWithDefault`.

Users can now also properly set match properties with the `-l` flag. Before these resulted in parsing errors since the Link CRD did not support these properly

```shell
$ bin/linkerd --context k3d-x multicluster link --cluster-name k3d-x -l 'mirror.linkerd.io/exported in (true)' |kubectl --context k3d-y apply -f -
...
$ bin/linkerd --context k3d-x mc link --cluster-name k3d-x --api-server-address https://172.18.0.3:6443 -l 'foo.bar=mirror-me' > link-x.yaml
```

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>