Allow AuthorizationPolicys to reference ServiceAccounts (#8709)
Closes #8565.

With this change, AuthorizationPolicys can now reference ServiceAccounts for
their target authentications. This allows users to avoid the requirement of
creating  a MeshTLSAuthentication resource that references a single
ServiceAccount.

The policy admission controller only allows an AuthorizationPolicy to reference
a single MeshTLSAuthentication _or_ a ServiceAccount; it cannot reference both.
Additionally, if a ServiceAccount is reference it can onl be a single
one—similar to MeshTLSAuthentications.

Signed-off-by: Kevin Leimkuhler <kleimkuhler@icloud.com>
Signed-off-by: Dmitrii Ermakov <demonihin@gmail.com>