Add action to lint GitHub workflows (#8737)
This change adds a linter for GitHub Actions workflows. In most cases,
Linkerd's workflows cannot be properly tested until they are triggered
in CI. This can lead to issues that are caught only during reviews (if
caught at all) or when the actions are triggered. Since running actions
can be an expensive operation (mostly in time spent), it would be useful
to have an automated way of checking our workflows and actions for
failures.

To lint our workflows, we introduce a new action to lint the workflow
templates and check the embeded scripts. The action itself will run only
when files are changed in the `.github/` directory.

Additionally, this change includes fixes to our embedded scripts, as
flagged by the action that has been introduced.

The change pins actionlint to v1.6.15.

Closes #8728

Signed-off-by: Matei David <matei@buoyant.io>
Co-authored-by: Oliver Gould <ver@buoyant.io>
Signed-off-by: Dmitrii Ermakov <demonihin@gmail.com>